unsignt int

+++ Board.cpp
@@ -30,12 +30,12 @@ std::string Board::read(unsigned int row, unsigned int column, Direction d , uns
 	//Check if in row and cols are in the allocated board
 	if(row+length<rows && column+length<cols){ 
 		if(d==Direction::Horizontal){	
-			for(int i=0;i<length;++i){
+			for(size_t i=0;i<length;++i){
 				str += board[row].at(i);
 			}
 		}
 		else{ //Vertical
-			for(int i=0;i<length;++i){
+			for(size_t i=0;i<length;++i){
 				str += board[i].at(column);
 			}
 		}
